refactor: make MCP kernel feature-neutral [ORB-10319]
