+++ crates/orbit-mcp/Cargo.toml
@@ -13,12 +13,9 @@ workspace = true
 
 [dependencies]
 orbit-common = { path = "../orbit-common" }
-orbit-graph = { path = "../orbit-graph" }
-orbit-graph-extract = { path = "../orbit-graph-extract" }
 rmcp = { version = "1.5", default-features = false, features = ["client", "server", "transport-io"] }
 serde.workspace = true
 serde_json.workspace = true
-sha2.workspace = true
 tokio.workspace = true
 tracing.workspace = true
