+++ crates/orbit-mcp/src/adapter/mod.rs
@@ -1,13 +1,9 @@
 //! The `dispatch` submodule owns MCP `tools/list` and `tools/call` wire framing and fans advertised tool names into host dispatch.
-//! The `graph` submodule owns direct orbit-graph MCP wrappers that live in this long-running process.
 //! The `structured` submodule owns the final `structuredContent` framing for strict MCP clients.
 //! The `schema` submodule emits JSON input schemas from Orbit tool metadata.
 //! The `name_map` submodule owns canonical-to-advertised tool name mapping and collision detection.
-//! The `learning_sidecar` submodule owns learning reminder lookup, session admission, and response sidecar injection.
 
 mod dispatch;
-mod graph;
-mod learning_sidecar;
 mod name_map;
 pub(crate) mod schema;
 mod structured;
@@ -21,15 +17,12 @@ mod tests;
 use std::collections::HashMap;
 use std::sync::{Arc, RwLock};
 
-#[cfg(test)]
-use orbit_common::types::{LearningInjectionCaps, LearningInjectionState};
-use orbit_common::types::{
-    McpToolDefinition, McpToolPolicyError, ToolParam, ToolSessionContext, audit_execution_id,
-};
+use orbit_common::types::{ToolParam, ToolSessionContext, audit_execution_id};
 
 use crate::{
-    McpCallContextResolver, McpCustomRequestHandler, McpHost, McpInputSchema, McpResultDecorator,
-    McpServerComposition, McpServerMetadata, McpToolExtension, McpToolExtensionRegistration,
+    McpCallContextResolver, McpCustomRequestHandler, McpHost, McpInputSchema,
+    McpInputSchemaResolver, McpResultDecorator, McpServerComposition, McpServerMetadata,
+    McpToolExtensionRegistration,
 };
 
 /// An rmcp [`ServerHandler`] that delegates tool listing and tool execution to
@@ -54,18 +47,20 @@ pub struct OrbitToolServer {
     extensions: Vec<McpToolExtensionRegistration>,
     result_decorators: Vec<Arc<dyn McpResultDecorator>>,
     call_context_resolver: Arc<dyn McpCallContextResolver>,
+    input_schema_resolver: Arc<dyn McpInputSchemaResolver>,
     custom_request_handlers: Vec<Arc<dyn McpCustomRequestHandler>>,
     metadata: McpServerMetadata,
     name_map: RwLock<HashMap<String, String>>,
     session_context: RwLock<ToolSessionContext>,
-    #[cfg(test)]
-    learning_states_for_test:
-        Option<Arc<tokio::sync::Mutex<std::collections::BTreeMap<String, LearningInjectionState>>>>,
 }
 
 impl OrbitToolServer {
     pub fn new(host: Arc<dyn McpHost>) -> Self {
-        Self::new_with_context(host, ToolSessionContext::trusted_local(None, None, None))
+        Self::new_with_context_and_composition(
+            host,
+            ToolSessionContext::trusted_local(None, None, None),
+            McpServerComposition::new(),
+        )
     }
 
     /// Construct a server with an explicit in-process extension composition.
@@ -90,8 +85,7 @@ impl OrbitToolServer {
     }
 
     pub fn new_with_context(host: Arc<dyn McpHost>, trusted_context: ToolSessionContext) -> Self {
-        let extensions = default_extensions(&host);
-        Self::new_with_context_and_extensions(host, trusted_context, extensions)
+        Self::new_with_context_and_composition(host, trusted_context, McpServerComposition::new())
     }
 
     /// Construct a server with trusted context and an explicit in-process
@@ -101,13 +95,7 @@ impl OrbitToolServer {
         trusted_context: ToolSessionContext,
         extensions: Vec<McpToolExtensionRegistration>,
     ) -> Self {
-        let composition = compatibility_composition(
-            &host,
-            extensions,
-            Arc::new(learning_sidecar::LearningSidecarDecorator::from_env(
-                Arc::clone(&host),
-            )),
-        );
+        let composition = McpServerComposition::new().with_tool_extensions(extensions);
         Self::new_with_context_and_composition(host, trusted_context, composition)
     }
 
@@ -130,86 +118,13 @@ impl OrbitToolServer {
             extensions: parts.tool_extensions,
             result_decorators: parts.result_decorators,
             call_context_resolver: parts.call_context_resolver,
+            input_schema_resolver: parts.input_schema_resolver,
             custom_request_handlers: parts.custom_request_handlers,
             metadata: parts.metadata,
             name_map: RwLock::new(HashMap::new()),
             session_context: RwLock::new(trusted_context),
-            #[cfg(test)]
-            learning_states_for_test: None,
         }
     }
-
-    #[cfg(test)]
-    fn new_for_test(
-        host: Arc<dyn McpHost>,
-        learning_session_id: Option<String>,
-        learning_caps: LearningInjectionCaps,
-        initial_state: LearningInjectionState,
-    ) -> Self {
-        let extensions = default_extensions(&host);
-        let learning = Arc::new(learning_sidecar::LearningSidecarDecorator::new_for_test(
-            Arc::clone(&host),
-            learning_session_id,
-            learning_caps,
-            initial_state,
-        ));
-        let learning_states = learning.states();
-        let composition = compatibility_composition(&host, extensions, learning);
-        let mut trusted_context = ToolSessionContext::trusted_local(None, None, None);
-        trusted_context.origin_session_id = Some(audit_execution_id("mcp-session"));
-        let mut server = Self::new_with_context_and_composition(host, trusted_context, composition);
-        server.learning_states_for_test = Some(learning_states);
-        server
-    }
-
-    #[cfg(test)]
-    async fn learning_state_for_test(&self, key: &str) -> Option<LearningInjectionState> {
-        let states = self.learning_states_for_test.as_ref()?;
-        states.lock().await.get(key).cloned()
-    }
-}
-
-fn compatibility_composition(
-    host: &Arc<dyn McpHost>,
-    extensions: Vec<McpToolExtensionRegistration>,
-    learning: Arc<dyn McpResultDecorator>,
-) -> McpServerComposition {
-    let mut composition = McpServerComposition::new()
-        .with_tool_extensions(extensions)
-        .with_result_decorator(learning)
-        .with_call_context_resolver(Arc::new(dispatch::CompatibilityCallContextResolver::new(
-            host.accepts_remote_session_context(),
-        )));
-    if host.accepts_remote_session_context() {
-        composition = composition.with_custom_request_handler(Arc::new(
-            dispatch::SpokeRegistrationHandler::new(Arc::clone(host)),
-        ));
-    }
-    if let Some(instructions) = host.private_server_instructions() {
-        composition =
-            composition.with_metadata(McpServerMetadata::default().with_instructions(instructions));
-    }
-    composition
-}
-
-fn default_extensions(host: &Arc<dyn McpHost>) -> Vec<McpToolExtensionRegistration> {
-    let graph: Arc<dyn McpToolExtension> = Arc::new(graph::GraphToolRegistry::new());
-    let registration = if host.in_process_graph_tools_enabled() {
-        McpToolExtensionRegistration::advertised(graph)
-    } else {
-        McpToolExtensionRegistration::recognition_only(graph)
-    };
-    vec![registration]
-}
-
-pub(super) const PROCESS_LEARNING_SESSION_KEY: &str = "__process__";
-
-pub(crate) fn graph_tool_names() -> &'static [&'static str] {
-    graph::GRAPH_TOOL_NAMES
-}
-
-pub(crate) fn graph_mcp_tool_definitions() -> Result<Vec<McpToolDefinition>, McpToolPolicyError> {
-    graph::graph_tool_definitions()
 }
 
 pub(crate) fn encode_mcp_input_schema(tool_name: &str, params: &[ToolParam]) -> McpInputSchema {
