+++ crates/orbit-mcp/src/adapter/schema.rs
@@ -13,7 +13,11 @@ pub(super) fn schema_to_tool(schema: ToolSchema, input_schema: JsonObject) -> To
 }
 
 pub(crate) fn build_input_schema(tool_name: &str, params: &[ToolParam]) -> JsonObject {
-    build_input_schema_with_enum_values(tool_name, params, enum_values_for)
+    build_input_schema_with_enum_values(tool_name, params, no_enum_values)
+}
+
+fn no_enum_values(_tool_name: &str, _param_name: &str) -> Option<&'static [&'static str]> {
+    None
 }
 
 pub(crate) fn build_input_schema_with_enum_values<F>(
@@ -65,50 +69,6 @@ where
     // schema validator.
     schema.insert("additionalProperties".to_string(), Value::Bool(true));
     schema
-}
-
-const TASK_TYPE_ENUM: &[&str] = &["feature", "bug", "refactor", "chore"];
-
-const TASK_UPDATE_STATUS_ENUM: &[&str] = &[
-    "proposed",
-    "backlog",
-    "someday",
-    "in-progress",
-    "review",
-    "done",
-    "blocked",
-    "rejected",
-];
-
-const TASK_COMPLEXITY_ENUM: &[&str] = &["low", "medium", "hard"];
-const AGENT_FAMILY_ENUM: &[&str] = &["codex", "claude", "gemini", "grok"];
-const GRAPH_SEARCH_KIND_ENUM: &[&str] = &["symbol", "string", "config"];
-const GRAPH_REF_CONFIDENCE_ENUM: &[&str] = &["exact", "import", "same_module", "fuzzy"];
-const GRAPH_REF_KIND_ENUM: &[&str] = &[
-    "call",
-    "type",
-    "use",
-    "trait_bound",
-    "impl",
-    "extends",
-    "implements",
-];
-
-pub(super) fn enum_values_for(
-    tool_name: &str,
-    param_name: &str,
-) -> Option<&'static [&'static str]> {
-    match (tool_name, param_name) {
-        ("orbit.task.add", "type") => Some(TASK_TYPE_ENUM),
-        ("orbit.task.update", "type") => Some(TASK_TYPE_ENUM),
-        ("orbit.task.update", "status") => Some(TASK_UPDATE_STATUS_ENUM),
-        ("orbit.task.add", "complexity") => Some(TASK_COMPLEXITY_ENUM),
-        ("orbit.graph.search", "kind") => Some(GRAPH_SEARCH_KIND_ENUM),
-        ("orbit.graph.refs", "confidence") => Some(GRAPH_REF_CONFIDENCE_ENUM),
-        ("orbit.graph.refs", "kind") => Some(GRAPH_REF_KIND_ENUM),
-        (_, "model") => Some(AGENT_FAMILY_ENUM),
-        _ => None,
-    }
 }
 
 /// Build the JSON-Schema fragment for a single parameter.
