+++ crates/orbit-mcp/src/adapter/tests/mod.rs
@@ -1,6 +1,4 @@
 mod dispatch;
-mod graph;
-mod learning_sidecar;
 mod name_map;
 mod schema;
 mod structured;
