+++ crates/orbit-remote/Cargo.toml
@@ -19,12 +19,15 @@ chrono.workspace = true
 hostname.workspace = true
 orbit-common = { path = "../orbit-common" }
 orbit-core = { path = "../orbit-core" }
+orbit-graph = { path = "../orbit-graph" }
+orbit-graph-extract = { path = "../orbit-graph-extract" }
 orbit-mcp = { path = "../orbit-mcp" }
 orbit-store = { path = "../orbit-store" }
 rusqlite.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_yaml.workspace = true
+sha2.workspace = true
 tokio = { workspace = true, features = ["io-util", "process", "time"] }
 toml.workspace = true
 tracing.workspace = true
@@ -33,5 +36,6 @@ tracing.workspace = true
 # Cross-layer compatibility canary: Remote MCP and the CLI learning hook must
 # share one persisted learning-admission state.
 orbit-cmd = { path = "../orbit-cmd" }
-rmcp = { version = "1.5", default-features = false, features = ["server", "transport-io"] }
+rmcp = { version = "1.5", default-features = false, features = ["client", "server", "transport-io"] }
 tempfile.workspace = true
+tracing-subscriber.workspace = true
