+++ crates/orbit-remote/src/mcp/mod.rs
@@ -1,10 +1,16 @@
 //! Remote MCP broker, coordination hub, and spoke-link composition.
 
 mod config;
+mod contract;
+mod graph;
 mod host;
 mod hub;
+mod hub_client;
 mod hub_link;
+mod learning;
 mod registration;
+mod schema;
+mod transport;
 
 #[cfg(test)]
 mod tests;
@@ -15,14 +21,21 @@ use std::sync::Arc;
 use orbit_common::types::{McpCapability, ToolSessionContext};
 use orbit_core::OrbitError;
 use orbit_core::runtime::resolve_global_root;
-use orbit_mcp::{McpHost, hub_schema_digest};
+use orbit_mcp::{
+    McpHost, McpResultDecorator, McpServerComposition, McpServerMetadata, McpToolExtension,
+    McpToolExtensionRegistration,
+};
 
 use crate::{HostIdentityState, HostMode, inspect_host_identity};
 
 use self::config::load_trusted_mcp_config;
+use self::contract::hub_schema_digest;
 use self::host::BrokerMcpHost;
 use self::hub::HubMcpHost;
 use self::hub_link::HubLinkPool;
+use self::learning::{LearningSidecarDecorator, LearningSidecarHost};
+use self::schema::RemoteInputSchemaResolver;
+use self::transport::{RemoteCallContextResolver, SpokeRegistrationHandler};
 
 pub use self::host::{canonical_mcp_tool_definitions, safe_mcp_tool_names};
 pub use self::registration::register_local_spoke;
@@ -45,17 +58,22 @@ pub fn serve_mcp_stdio(
         ));
     }
     let capability = requested_capability.unwrap_or(McpCapability::Agent);
-    let (host, mut trusted_context): (Arc<dyn McpHost>, ToolSessionContext) = if hub {
-        let hub = HubMcpHost::new(global_root.clone(), capability)?;
+    let (host, mut trusted_context, composition): (
+        Arc<dyn McpHost>,
+        ToolSessionContext,
+        McpServerComposition,
+    ) = if hub {
+        let hub = Arc::new(HubMcpHost::new(global_root.clone(), capability)?);
         let identity = hub.identity();
         let context = ToolSessionContext::trusted_local(
             None,
             Some(identity.machine_id.clone()),
             Some(identity.host_id.clone()),
         );
-        (Arc::new(hub), context)
+        let composition = hub_server_composition(Arc::clone(&hub));
+        (hub, context, composition)
     } else {
-        let (host, machine_id, host_id): (Arc<dyn McpHost>, Option<String>, Option<String>) =
+        let (host, machine_id, host_id): (Arc<BrokerMcpHost>, Option<String>, Option<String>) =
             match inspect_host_identity(&global_root)? {
                 HostIdentityState::Present(identity) => {
                     if identity.mode == HostMode::Spoke {
@@ -92,8 +110,9 @@ pub fn serve_mcp_stdio(
                 ),
             };
         (
-            host,
+            Arc::clone(&host) as Arc<dyn McpHost>,
             ToolSessionContext::trusted_local(None, machine_id, host_id),
+            broker_server_composition(host),
         )
     };
     trusted_context.effective_capabilities = BTreeSet::from([capability]);
@@ -102,5 +121,34 @@ pub fn serve_mcp_stdio(
         .enable_all()
         .build()
         .map_err(|error| OrbitError::Execution(format!("tokio runtime: {error}")))?;
-    tokio_runtime.block_on(orbit_mcp::serve_stdio_with_context(host, trusted_context))
+    tokio_runtime.block_on(orbit_mcp::serve_stdio_with_context_and_composition(
+        host,
+        trusted_context,
+        composition,
+    ))
+}
+
+fn broker_server_composition(host: Arc<BrokerMcpHost>) -> McpServerComposition {
+    let graph: Arc<dyn McpToolExtension> = Arc::new(graph::GraphToolRegistry::new());
+    let learning_host: Arc<dyn LearningSidecarHost> = host;
+    let learning: Arc<dyn McpResultDecorator> =
+        Arc::new(LearningSidecarDecorator::from_env(learning_host));
+    McpServerComposition::new()
+        .with_tool_extension(McpToolExtensionRegistration::advertised(graph))
+        .with_result_decorator(learning)
+        .with_input_schema_resolver(Arc::new(RemoteInputSchemaResolver))
+}
+
+fn hub_server_composition(host: Arc<HubMcpHost>) -> McpServerComposition {
+    let graph: Arc<dyn McpToolExtension> = Arc::new(graph::GraphToolRegistry::new());
+    let learning_host: Arc<dyn LearningSidecarHost> = host.clone();
+    let learning: Arc<dyn McpResultDecorator> =
+        Arc::new(LearningSidecarDecorator::from_env(learning_host));
+    McpServerComposition::new()
+        .with_tool_extension(McpToolExtensionRegistration::recognition_only(graph))
+        .with_result_decorator(learning)
+        .with_call_context_resolver(Arc::new(RemoteCallContextResolver))
+        .with_input_schema_resolver(Arc::new(RemoteInputSchemaResolver))
+        .with_custom_request_handler(Arc::new(SpokeRegistrationHandler::new(Arc::clone(&host))))
+        .with_metadata(McpServerMetadata::default().with_instructions(host.private_instructions()))
 }
