+++ crates/orbit-remote/src/mcp/tests/snapshots/wire_tools_list.json
