+++ docs/design-patterns/test_layout.md
@@ -88,7 +88,7 @@ This nests the test module as a *child* of the source module. Children can read 
 
 ## Reference: `orbit-mcp::adapter`
 
-The `adapter` module has children `name_map.rs`, `schema.rs`, `dispatch.rs`, `structured.rs`, `learning_sidecar.rs`. The canonical layout: `adapter/mod.rs` declares each child plus `#[cfg(test)] mod tests;`, and `adapter/tests/` contains one file per source child (`tests/name_map.rs`, `tests/schema.rs`, etc.). Each test file accesses its sibling source via `use super::super::<sibling_module>::<item>;`.
+The generic MCP `adapter` module has children `name_map.rs`, `schema.rs`, `dispatch.rs`, and `structured.rs`. The canonical layout: `adapter/mod.rs` declares each child plus `#[cfg(test)] mod tests;`, and `adapter/tests/` contains one file per source child (`tests/name_map.rs`, `tests/schema.rs`, etc.). Remote-owned graph and learning composition tests live beside their implementations under `orbit-remote/src/mcp/tests/`. Each test file accesses its sibling source via `use super::super::<sibling_module>::<item>;`.
 
 ## Migration recipe
