+++ docs/design/orbit-graph/2_design.md
@@ -195,7 +195,7 @@ Bounded outputs: `impact` and `trace` both cap at 200 visited nodes regardless o
 
 `refs` unions queries against the `refs` table (calls/type/use/trait_bound) and the `relations` table (impl/extends/implements). CLI `--kind impl` is a routing alias to `relations`. The ten-command surface absorbs `orbit-knowledge`'s navigation use cases while keeping `overview`, `implementors`, and outbound module-level `deps` explicit.
 
-For the agent-facing MCP entry point, the ten canonical names are an explicit safe-surface set in `orbit-cli`. `orbit-mcp` owns the in-process implementations but calls them only through the host's policy/audit seam; `RuntimeMcpHost` applies the allowlist and records success or failure through the shared `ToolEntryPoint::Mcp` runtime boundary. Known graph schemas accidentally re-exposed by a host are replaced by the adapter schemas, so schema presence cannot disable that seam. [ORB-10225] added the startup subset assertion and end-to-end audit coverage.
+For the agent-facing MCP entry point, the ten canonical names are an explicit safe-surface set in `orbit-remote`. `orbit-remote` owns and explicitly composes the in-process implementations over the generic `orbit-mcp` kernel, but calls them only through the host's policy/audit seam; the broker host applies the allowlist and records success or failure through the shared `ToolEntryPoint::Mcp` runtime boundary. Known graph schemas accidentally re-exposed by a host are replaced by the Remote extension schemas, so schema presence cannot disable that seam. [ORB-10225] added the startup subset assertion and end-to-end audit coverage.
 
 The `Selector` grammar is preserved verbatim from `orbit-knowledge` — every form used in existing skills (`symbol:<file>#<name>:<kind>`, `file:<path>`, `module:<qualified>`, `command:<name>`) must continue to parse identically. A pre-Step-1 audit of `.claude/skills/` captures the full grammar surface as the canonical reference.
