+++ docs/design/orbit-graph/4_decisions.md
@@ -158,11 +158,11 @@ Format for each entry: **Status · Date · Task(s)**, then *Context → Decision
 
 **Context.** ADR-0192 rolled back the ORB-00338 v2 cutover and restored `orbit-knowledge` (v1) as the active graph backend, gating any future cutover on a set of pre-cutover correctness/latency learnings. ORB-00386/00387/00389/00390 closed those gates against the watcher-backed orbit-graph (v2): UTF-8 `show` boundaries, trace/impact correctness, the `overview`/`implementors`/`deps` navigation queries (restoring v1 parity), and the `refs` fuzzy fallback. The automated equivalence/effectiveness harness that ADR-0192/§16 envisioned was removed by ADR-0197 and never rebuilt; with the gates closed, rebuilding it as a precondition carried no proportionate value.
 
-**Decision.** Cut the agent graph surface over to orbit-graph (v2) and remove `orbit-knowledge` entirely. The v2 tools (`sync`, `search`, `show`, `refs`, `callees`, `impact`, `trace`, `overview`, `implementors`, `deps`) are served by the in-process `GraphToolRegistry` in `orbit-mcp`, which activates whenever the host exposes no `orbit.graph.*` schema. The v1 builtins, the `orbit graph` CLI command, the init-time graph build, and the v1 metrics pipeline are removed; the knowledge-stats computation moves to `orbit_core::metrics`. The measurement bar for this cutover is **manual QA plus a v1-vs-v2 spot-check**, accepted in place of the never-rebuilt automated harness. The `pack` tool is dropped rather than ported (ORB-00388 rejected); `callers` is subsumed by `refs`.
+**Decision.** Cut the agent graph surface over to orbit-graph (v2) and remove `orbit-knowledge` entirely. The v2 tools (`sync`, `search`, `show`, `refs`, `callees`, `impact`, `trace`, `overview`, `implementors`, `deps`) are served by the in-process `GraphToolRegistry` explicitly composed by `orbit-remote` over the generic `orbit-mcp` kernel. The v1 builtins, the `orbit graph` CLI command, the init-time graph build, and the v1 metrics pipeline are removed; the knowledge-stats computation moves to `orbit_core::metrics`. The measurement bar for this cutover is **manual QA plus a v1-vs-v2 spot-check**, accepted in place of the never-rebuilt automated harness. The `pack` tool is dropped rather than ported (ORB-00388 rejected); `callers` is subsumed by `refs`.
 
 **Consequences.**
 - orbit-graph is the sole graph surface; `orbit-knowledge` is deleted from the workspace (~24k LOC, plus the v1 builtins and the `orbit graph` CLI command).
-- Agents reach the graph only through `orbit-mcp`; there is no `orbit graph` subcommand. The standalone `orbit-graph-cli` binary remains for direct CLI use. *(Amended by ADR-0199: `orbit graph` is reintroduced as a thin wrapper over the `orbit-graph-cli` library for human/script use; the agent surface stays MCP-only.)*
+- Agents reach the graph only through the Remote-composed MCP surface; there is no `orbit graph` subcommand. The standalone `orbit-graph-cli` binary remains for direct CLI use. *(Amended by ADR-0199: `orbit graph` is reintroduced as a thin wrapper over the `orbit-graph-cli` library for human/script use; the agent surface stays MCP-only.)*
 - The dashboard knowledge-stats panel keeps working via `orbit_core::metrics::aggregate`; pack-compression fields degrade to defaults (pack is gone), matching the pre-existing pack-less code path.
 - ADR-0192's rollback decision is reversed; its `SyncPolicy::Manual` default note is superseded by the watcher-backed policy (ADR-0195) for long-lived MCP handles.
 - Cost: the v1 content-addressed store, working-graph write surface, and graph-history attribution are gone; there is no automated equivalence gate guarding regressions against the (now-removed) v1 backend.
