+++ docs/design/orbit-graph/specs/GRAPH_SPEC.md
@@ -397,7 +397,7 @@ orbit graph implementors <trait-selector>
 orbit graph deps <file:… | dir:…>
 ```
 
-The agent-facing MCP surface uses these same ten canonical names as an explicit `orbit-cli` safe-surface allowlist. Although `orbit-mcp` executes the graph implementations in-process to retain watcher-backed handles, every call first crosses the host-owned policy/audit seam. The production host performs allowlist preflight inside OrbitRuntime's shared `ToolEntryPoint::Mcp` audit bracket, recording tool name, role, duration, and success/failure; a rejected name never invokes the graph handler. The MCP host asserts at startup that the adapter name set is a subset of the intended safe surface. If a host re-exposes a known graph schema, the adapter schema and policy/audit route remain authoritative rather than falling back based on schema presence. [ORB-10225]
+The agent-facing MCP surface uses these same ten canonical names as an explicit `orbit-remote` safe-surface allowlist. `orbit-remote` executes the graph implementations in-process over the generic `orbit-mcp` kernel to retain watcher-backed handles, and every call first crosses the host-owned policy/audit seam. The production host performs allowlist preflight inside OrbitRuntime's shared `ToolEntryPoint::Mcp` audit bracket, recording tool name, role, duration, and success/failure; a rejected name never invokes the graph handler. The broker host asserts at startup that the extension name set is a subset of the intended safe surface. If a host re-exposes a known graph schema, the Remote extension schema and policy/audit route remain authoritative rather than falling back based on schema presence. [ORB-10225]
 
 ### 9.1 `search`
 
@@ -632,7 +632,7 @@ Estimated landing: ~24k → ~10k LOC. More capability (string / command / config
 
 ## 16. Migration plan
 
-> **Status — completed (ORB-00391, 2026-06).** The migration is done: `orbit-graph` (v2) is the sole graph surface and the `orbit-knowledge` (v1) crate has been removed. The agent-facing `orbit.graph.*` tools are served by the in-process orbit-graph adapter in `orbit-mcp`; the v1 builtins, the `orbit graph` CLI command, the init-time graph build, and the v1 metrics pipeline were decommissioned (the knowledge-stats computation moved to `orbit_core::metrics`). Step 4's automated effectiveness/equivalence harness was never rebuilt after ADR-0197 removed it; the accepted measurement bar for the final cutover was **manual QA plus a v1-vs-v2 spot-check**, not the harness described below. See ADR-0192 (superseded by ADR-0198). The four-step plan below is retained as the historical design record.
+> **Status — completed (ORB-00391, 2026-06).** The migration is done: `orbit-graph` (v2) is the sole graph surface and the `orbit-knowledge` (v1) crate has been removed. The agent-facing `orbit.graph.*` tools are served by the in-process orbit-graph adapter in `orbit-remote` over the generic MCP kernel; the v1 builtins, the `orbit graph` CLI command, the init-time graph build, and the v1 metrics pipeline were decommissioned (the knowledge-stats computation moved to `orbit_core::metrics`). Step 4's automated effectiveness/equivalence harness was never rebuilt after ADR-0197 removed it; the accepted measurement bar for the final cutover was **manual QA plus a v1-vs-v2 spot-check**, not the harness described below. See ADR-0192 (superseded by ADR-0198). The four-step plan below is retained as the historical design record.
 
 A four-step Orbit epic. Each step is one or more tasks; each task is independently shippable.
