+++ docs/design/project-learnings/references/glossary.md
@@ -10,7 +10,7 @@ Project-specific vocabulary used in [1_overview.md](../1_overview.md), [2_design
 | **Engine pre-prompt injection** | Layer 1 of the push pipeline. `orbit-engine` queries matching learnings before spawning an agent and prepends summaries to the agent prompt. Universal across agents. See [2_design.md §4.1](../2_design.md). |
 | **Evidence** | Provenance attached to a learning record — commit SHAs, task IDs, or external refs that produced or substantiate the learning. See [2_design.md §2.1](../2_design.md). |
 | **Learning record** | The first-class Orbit resource representing one piece of project knowledge. YAML on disk, indexed in SQLite, mutated through `orbit.learning.*` tools. See [2_design.md §2](../2_design.md). |
-| **MCP-sidecar injection** | Layer 2 of the push pipeline. `orbit-mcp` attaches a `learnings` field to MCP tool responses whose arguments or output reference file paths. Cross-agent. See [2_design.md §4.2](../2_design.md). |
+| **MCP-sidecar injection** | Layer 2 of the push pipeline. `orbit-remote` composes a decorator that attaches a `learnings` field to MCP tool responses whose arguments or output reference file paths. Cross-agent. See [2_design.md §4.2](../2_design.md). |
 | **Pre-tool-use hook** | Layer 3 of the push pipeline. A Claude Code `PreToolUse` hook on `Edit | Write | Read` injects matching learnings before the tool fires. Claude-Code-only. See [2_design.md §4.3](../2_design.md). |
 | **Pull surface** | The `orbit.search` tool (with `kind: "learning"`) and the `orbit-learnings` skill. Used for active query at task start or when an agent has time to ask. Complement to push, not substitute. See [2_design.md §6](../2_design.md). |
 | **Push layer** | One of the three injection sites that surface learnings into agent context without an agent query. The three layers are engine pre-prompt, MCP-sidecar, and Claude Code hook. See [2_design.md §4](../2_design.md). |
